Build against system Armadillo, fix a crash on the shipped data, add docs
The repository did not run out of the box: all three example scripts pointed at
input files that are not in it, and the one combination that does match the
shipped data crashed inside the sequence-similarity reader.

Crash:
- readSeqSim_SMAT read its input as SMAT without checking for the "rows cols
  nnz" header. Handed a headerless triple list, it took the first data row as
  the header, allocated from those numbers, and then wrote at w[n*row+col] with
  row and col still -1. The reader now validates the header, checks every entry
  against the declared dimensions, reports the line number, and suggests -t tab
  when the file looks label-keyed.
- mem_size was an int in both readSeqSim_SMAT and the uniform-prior fallback in
  main.cc. rows * cols * sizeof(double) overflows 2^31 for interactomes only a
  little larger than the ones shipped here. Both are size_t now.

Example scripts:
- sparse_run.sh, dense_run.sh, and dense_run_noPrior.sh all referenced
  BioGRID_full_*_net.tsv, which does not exist; the repository ships .smat.
  They also paired those with the .evals similarity file, which is keyed by
  Entrez gene ID and belongs to the tab reader, not the SMAT reader. All three
  now use -t smat with BioGRID_full_human-yeast.smat, and each carries a comment
  explaining what it demonstrates.

Build:
- Remove the vendored Armadillo 5.200: 926 files and 19 MB, including one
  developer's CMake build tree with an absolute-path CMakeCache.txt, compiled
  object files, and an a.out. TAME now links against the system Armadillo,
  which the README documents per platform.
- Rewrite the Makefile: C++17 (required by Armadillo 6 and later), portable
  -O2 in place of -O4 -march=native -msse2 -fforce-addr, honour
  CXX/CXXFLAGS/LDLIBS/PREFIX, correct the bMatch rules (the pattern matched
  .cc while the sources are .cpp, and Node.cpp was passed straight to the
  linker), and add check, check-badinput, and install targets.
- Remove the Eclipse .cproject, .project, and .settings/ workspace metadata.

Tests and CI:
- Add tests/data: a small synthetic network pair and similarity prior, so
  `make check` verifies the pipeline in seconds instead of running a full
  interactome alignment.
- Add `make check-badinput`, a regression test for the crash above: it asserts
  TAME rejects a headerless similarity file and does not die on a signal.
- Add a GitHub Actions workflow building and testing on Linux with GCC and
  Clang and on macOS.

Documentation:
- Rewrite README.md: what higher-order alignment is and why triangles, the
  pipeline, per-platform dependency install, build and install, the full
  option list, all three input formats, what each bundled file is and which
  reader it belongs to, worked examples, and the citation.
- Add LICENSE (MIT), CITATION.cff, and .gitignore.

+++ src/main.cc
@@ -363,8 +363,8 @@ int main(int argc, char **argv) {
 	}
 	if( w == NULL) {
 		fprintf(stderr, "Main:: Error reading sequence similarity file. Resetting to uniform prior\n");
-		long N = (G->n*H->n);
-		int mem_size = N*sizeof(double);
+		long N = ((long)G->n * (long)H->n);
+		size_t mem_size = (size_t)N * sizeof(double);
 		w = (double *)calloc(1, mem_size);
 		if(w == NULL) {
 			fprintf(stderr, "main:: Error allocating memory for w.\n");
